+++ dev_scripts_helpers/notebooks/run_jupyter_server.py
@@ -77,7 +77,7 @@ def _start(port, action):
     print(f"You can connect to: {ip_name}:{port}")
     cmd = f"jupyter notebook '--ip=*' --browser chrome . --port {port}"
     if action != "only_print_cmd":
-        hsystem.system(cmd, suppress_output=False, log_level="echo")
+        hsystem.system(cmd, suppress_output=False, log_level="PRINT")
     else:
         print(cmd)
