+++ helpers/hdbg.py
@@ -851,6 +851,32 @@ def dassert_dir_exists(
         _dfatal(txt, msg, *args, only_warning=only_warning)
 
 
+def dassert_valid_path(file_path: str, is_input: bool) -> None:
+    """
+    Assert that a file path is valid, based on it being input or output.
+
+    For input files, it ensures that the file or directory exists. For
+    output files, it ensures that the enclosing directory exists.
+
+    :param file_path: The file path to check.
+    :param is_input: Whether the file path is an input file.
+    """
+    if is_input:
+        # If it's an input file, then `file_path` must exist as a file or a dir.
+        dassert_path_exists(file_path)
+    else:
+        # If it's an output, we might be writing a file that doesn't exist yet,
+        # but we assume that the including directory is already present.
+        dir_name = os.path.normpath(os.path.dirname(file_path))
+        os.makedirs(dir_name, exist_ok=True)
+        dassert(
+            os.path.exists(file_path) or os.path.exists(dir_name),
+            "Invalid path: '%s' and '%s' don't exist",
+            file_path,
+            dir_name,
+        )
+
+
 def dassert_file_extension(
     file_name: str,
     extensions: Union[str, List[str]],
